LC-335 Split GUI file
